+++ src/components/useThree/mainCanvas.tsx
@@ -9,13 +9,13 @@ export const MainCanvas = () => {
         fov: 45,
         near: 0.1,
         far: 1000,
-        position: [0, 0, 5],
+        position: [0, 0, 0],
       }}
       style={{ background: "black" }}
     >
       <directionalLight position={[1, 1, 1]} intensity={0.8} />
       <ambientLight args={[0xffffff]} intensity={0.5} />
-      <OrbitControls minDistance={15} maxDistance={70} />
+      <OrbitControls minDistance={250} maxDistance={500} />
       <Moon />
     </Canvas>
   );
